+++ do.go
@@ -10,7 +10,6 @@ import (
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
 	"gorm.io/gorm/schema"
-	"gorm.io/hints"
 
 	"gorm.io/gen/field"
 )
@@ -137,44 +136,6 @@ func (d *DO) Clauses(conds ...clause.Expression) Dao {
 		return d.getInstance(newDB)
 	}
 	return d.getInstance(d.db.Clauses(conds...))
-}
-
-func checkConds(conds []clause.Expression) error {
-	for _, cond := range conds {
-		if err := checkClause(cond); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-var banClauses = map[string]bool{
-	"INSERT":      true,
-	"VALUES":      true,
-	"ON CONFLICT": true,
-	"SELECT":      true,
-	"FROM":        true,
-	"WHERE":       true,
-	"GROUP BY":    true,
-	"ORDER BY":    true,
-	"LIMIT":       true,
-	"FOR":         true,
-	"UPDATE":      true,
-	"SET":         true,
-	"DELETE":      true,
-}
-
-func checkClause(cond clause.Expression) error {
-	switch cond := cond.(type) {
-	case hints.Hints, hints.IndexHint:
-		return nil
-	case clause.Interface:
-		if banClauses[cond.Name()] {
-			return fmt.Errorf("banned clause %s", cond.Name())
-		}
-		return nil
-	}
-	return fmt.Errorf("unknown clause %v", cond)
 }
 
 // As alias cannot be heired, As must used on tail
